requirements.txt
daily-python
deepgram-sdk>=3.0.0
elevenlabs>=1.0.0
openai>=1.0.0
anthropic>=0.20.0
google-generativeai>=0.4.0
flask>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
